+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 0.8.0 - 2026-07-29
 ### Fixed
 - Bug which breaks sites using Akamai Bot Manager 
 - Ensure Top 10 alert tags have expected URLs.